Fix projection comparison
Using the length of the projection vector meant that polygons in negative
coordinates could be interpreted as overlapping polygons in the same positive
coordinates. Use the the x or y position instead, whichever's magnitude is
greater. This avoids the case where the edge being projected onto is completely
horizontal or vertical, which would cause all y or x values to be 0.